requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
langchain-core
sentence-transformers
numpy
fastapi
python-dotenv
uvicorn
docarray
